INSPECTIT-1986: Connection error to a CMR in the UI shouldn't have a popup
