+++ inspectit.ui.rcp/src/main/java/rocks/inspectit/ui/rcp/repository/CmrRepositoryManager.java
@@ -8,15 +8,9 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 import org.apache.commons.collections.CollectionUtils;
-import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.core.runtime.IStatus;
-import org.eclipse.core.runtime.Status;
-import org.eclipse.core.runtime.jobs.Job;
-import org.eclipse.ui.progress.IProgressConstants;
 
 import rocks.inspectit.shared.all.cmr.model.PlatformIdent;
-import rocks.inspectit.ui.rcp.InspectIT;
-import rocks.inspectit.ui.rcp.InspectITImages;
+import rocks.inspectit.ui.rcp.job.UpdateRepositoryJob;
 import rocks.inspectit.ui.rcp.preferences.PreferencesUtils;
 import rocks.inspectit.ui.rcp.util.ListenerList;
 
@@ -242,63 +236,4 @@ public class CmrRepositoryManager {
 		}
 		PreferencesUtils.saveCmrRepositoryDefinitions(toSave, false);
 	}
-
-	/**
-	 * Update online status of all repositories job.
-	 *
-	 * @author Ivan Senic
-	 *
-	 */
-	public static class UpdateRepositoryJob extends Job {
-
-		/**
-		 * CMR to update.
-		 */
-		private CmrRepositoryDefinition cmrRepositoryDefinition;
-
-		/**
-		 * Should job be rescheduled after its execution.
-		 */
-		private boolean rescheduleJob;
-
-		/**
-		 * Default constructor.
-		 *
-		 * @param cmrRepositoryDefinition
-		 *            {@link CmrRepositoryDefinition} to update.
-		 * @param rescheduleJob
-		 *            If job should be rescheduled after execution.
-		 */
-		public UpdateRepositoryJob(CmrRepositoryDefinition cmrRepositoryDefinition, boolean rescheduleJob) {
-			super("Updating online status of CMR repository " + cmrRepositoryDefinition.getIp() + ":" + cmrRepositoryDefinition.getPort());
-			this.cmrRepositoryDefinition = cmrRepositoryDefinition;
-			this.rescheduleJob = rescheduleJob;
-			this.setUser(false);
-			this.setProperty(IProgressConstants.ICON_PROPERTY, InspectIT.getDefault().getImageDescriptor(InspectITImages.IMG_SERVER_REFRESH_SMALL));
-		}
-
-		/**
-		 * {@inheritDoc}
-		 */
-		@Override
-		protected IStatus run(IProgressMonitor monitor) {
-			try {
-				cmrRepositoryDefinition.refreshOnlineStatus();
-				return Status.OK_STATUS;
-			} finally {
-				if (rescheduleJob) {
-					this.schedule(UPDATE_JOB_REPETITION);
-				}
-			}
-		}
-
-		/**
-		 * @return the cmrRepositoryDefinition
-		 */
-		public CmrRepositoryDefinition getCmrRepositoryDefinition() {
-			return cmrRepositoryDefinition;
-		}
-
-	}
-
 }
