+++ inspectit.ui.rcp/src/main/java/rocks/inspectit/ui/rcp/view/impl/DataExplorerView.java
@@ -53,13 +53,13 @@ import rocks.inspectit.ui.rcp.InspectITImages;
 import rocks.inspectit.ui.rcp.editor.tree.DeferredTreeViewer;
 import rocks.inspectit.ui.rcp.formatter.ImageFormatter;
 import rocks.inspectit.ui.rcp.formatter.TextFormatter;
+import rocks.inspectit.ui.rcp.job.UpdateRepositoryJob;
 import rocks.inspectit.ui.rcp.model.TreeModelManager;
 import rocks.inspectit.ui.rcp.preferences.PreferencesConstants;
 import rocks.inspectit.ui.rcp.preferences.PreferencesUtils;
 import rocks.inspectit.ui.rcp.repository.CmrRepositoryChangeListener;
 import rocks.inspectit.ui.rcp.repository.CmrRepositoryDefinition;
 import rocks.inspectit.ui.rcp.repository.CmrRepositoryDefinition.OnlineStatus;
-import rocks.inspectit.ui.rcp.repository.CmrRepositoryManager.UpdateRepositoryJob;
 import rocks.inspectit.ui.rcp.repository.RepositoryDefinition;
 import rocks.inspectit.ui.rcp.repository.StorageRepositoryDefinition;
 import rocks.inspectit.ui.rcp.storage.listener.StorageChangeListener;
