+++ inspectit.ui.rcp/src/main/java/rocks/inspectit/ui/rcp/view/impl/RepositoryManagerView.java
@@ -70,6 +70,7 @@ import rocks.inspectit.ui.rcp.form.CmrRepositoryPropertyForm;
 import rocks.inspectit.ui.rcp.formatter.ImageFormatter;
 import rocks.inspectit.ui.rcp.formatter.TextFormatter;
 import rocks.inspectit.ui.rcp.handlers.ShowRepositoryHandler;
+import rocks.inspectit.ui.rcp.job.UpdateRepositoryJob;
 import rocks.inspectit.ui.rcp.model.AgentLeaf;
 import rocks.inspectit.ui.rcp.model.Component;
 import rocks.inspectit.ui.rcp.model.DeferredAgentsComposite;
@@ -78,7 +79,6 @@ import rocks.inspectit.ui.rcp.repository.CmrRepositoryChangeListener;
 import rocks.inspectit.ui.rcp.repository.CmrRepositoryDefinition;
 import rocks.inspectit.ui.rcp.repository.CmrRepositoryDefinition.OnlineStatus;
 import rocks.inspectit.ui.rcp.repository.CmrRepositoryManager;
-import rocks.inspectit.ui.rcp.repository.CmrRepositoryManager.UpdateRepositoryJob;
 import rocks.inspectit.ui.rcp.repository.RepositoryDefinition;
 import rocks.inspectit.ui.rcp.util.SafeExecutor;
 import rocks.inspectit.ui.rcp.view.IRefreshableView;
