+++ inspectit.ui.rcp/src/main/resources/META-INF/spring/spring-context-model-main.xml
@@ -109,6 +109,7 @@
 	<bean id="jmxDataAccessServiceTarget" class="rocks.inspectit.ui.rcp.repository.service.cmr.CmrService" scope="prototype" lazy-init="true">
 		<property name="serviceInterface" value="rocks.inspectit.shared.cs.cmr.service.IJmxDataAccessService" />
 		<property name="serviceName" value="JmxDataAccessService" />
+		<property name="defaultValueOnError" value="true" />
 	</bean>
 	<!-- JmxDataAccessService End -->
 
@@ -299,6 +300,7 @@
 	<bean id="businessContextManagementServiceTarget" class="rocks.inspectit.ui.rcp.repository.service.cmr.CmrService" scope="prototype" lazy-init="true">
 		<property name="serviceInterface" value="rocks.inspectit.shared.cs.cmr.service.IBusinessContextManagementService" />
 		<property name="serviceName" value="BusinessContextManagementService" />
+		<property name="defaultValueOnError" value="true" />
 	</bean>
 	<!-- BusinessContextManagementService End -->
 	
@@ -346,6 +348,7 @@
 	<bean id="influxDBServiceTarget" class="rocks.inspectit.ui.rcp.repository.service.cmr.CmrService" scope="prototype" lazy-init="true">
 		<property name="serviceInterface" value="rocks.inspectit.shared.cs.cmr.service.IInfluxDBService" />
 		<property name="serviceName" value="InfluxDBService" />
+		<property name="defaultValueOnError" value="true" />
 	</bean>
 	<!-- InfluxDBService End -->
 	
@@ -369,6 +372,7 @@
 	<bean id="alertAccessServiceTarget" class="rocks.inspectit.ui.rcp.repository.service.cmr.CmrService" scope="prototype" lazy-init="true">
 		<property name="serviceInterface" value="rocks.inspectit.shared.cs.cmr.service.IAlertService" />
 		<property name="serviceName" value="AlertAccessService" />
+		<property name="defaultValueOnError" value="true" />
 	</bean>
 	<!-- AlertAccessService End -->
 	
@@ -392,6 +396,7 @@
 	<bean id="agentInstrumentationServiceTarget" class="rocks.inspectit.ui.rcp.repository.service.cmr.CmrService" scope="prototype" lazy-init="true">
 		<property name="serviceInterface" value="rocks.inspectit.shared.cs.cmr.service.IAgentInstrumentationService" />
 		<property name="serviceName" value="AgentInstrumentationService" />
+		<property name="defaultValueOnError" value="true" />
 	</bean>
 	<!-- AlertAccessService End -->
 	
@@ -415,6 +420,7 @@
 	<bean id="spanServiceTarget" class="rocks.inspectit.ui.rcp.repository.service.cmr.CmrService" scope="prototype" lazy-init="true">
 		<property name="serviceInterface" value="rocks.inspectit.shared.cs.cmr.service.ISpanService" />
 		<property name="serviceName" value="SpanService" />
+		<property name="defaultValueOnError" value="true" />
 	</bean>
 	<!-- SpanService End -->
